% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocess.R
\name{data_preprocess}
\alias{data_preprocess}
\title{Outputs a synthetic survey using a simple model}
\usage{
data_preprocess(data, limits = NULL, dummycode = NULL)
}
\arguments{
\item{data}{The number of rows in the survey}

\item{limits}{The number of columns in the survey}

\item{dummycode}{The fraction of nodes in the smaller of the two polarised groups}
}
\value{
A data frame corresponding to a survey.
}
\description{
\code{data_preprocess()} outputs a synthetic survey, generated using a simple, stochastic
model of polarisation.
}
\examples{
S <- make_synthetic_data(200, 8)
}
