% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{make_projection}
\alias{make_projection}
\title{Outputs the survey projection onto the agent or symbolic layer}
\usage{
make_projection(
  data,
  layer,
  threshold_method = NULL,
  method_value = NULL,
  centre = NULL,
  similarity_metric = NULL
)
}
\arguments{
\item{data}{A data frame corresponding to a survey}

\item{layer}{A string flag specifying which layer to project
\itemize{
\item \code{"agent"} produces the network corresponding to the agents, which we assume
to be rows in the \code{data}
\item \code{"symbolic"} produces the network corresponding to the symbols, or items,
which we assume to be columns in the `data
}

If a string is provided that doesn't match the above two options, we default
to the agent layer.}

\item{threshold_method}{A string flag that specifies how edges are selected in
the network representation.
\itemize{
\item \code{"raw_similarity"} means we remove all edges whose weight (i.e. node
similarity) is below a specified threshold.
\item \code{"target_lcc"} finds the value of the threshold that results in the network
whose largest connected component (lcc) is as close as possible to a
specified value. In general, there will be a range of thresholds that satisfy
this condition, and as such, we choose the upper limit of this range. That
is, this method produces the sparsest possible network of all those with an
optimal lcc.
\item \code{"target_ad"} finds the value of the threshold that results in the network
whose average degree (ad) is as close as possible to that specified by
\code{method_value}
}

If a string is provided that doesn't match these options, the method defaults to
\code{"target_lcc"} with \code{threshold_method = 0.98}.}

\item{method_value}{A utility variable that we interpret according to the
\code{threshold_method} chosen.
\itemize{
\item If \code{threshold_method = "raw_similarity"}, then \code{method_value} is
interpreted as the similarity threshold, and thus is in the range \verb{[-1, 1]}.
A value of -1 means no edges are removed, and a value of 1 means all edges
are removed.
\item If \code{threshold_method = "target_lcc"}, then \code{method_value} is interpreted as
the desired fractional size of the largest connected component, in the range
\verb{[0, 1]}. E.g., when set to 0, no nodes are connected, and if set to 1, the
network is as sparse as possible while remaining fully connected.
\item If \code{threshold_method = "target_ad"}, then \code{method_value} is interpreted as
the desired average degree. We assume that \code{method_value} is normalised to
the range \verb{[0, 1]} When \code{method_value = 0}, then no nodes are connected, and
if \code{method_value = 1}, the network is complete, meaning it contains every
possible edge.
}}
}
\value{
A data.frame corresponding to the edge list of the specified network. It
contains three columns named \code{u}, \code{v} and \code{weight}, such that each row
represents an undirected, weighted edge.
}
\description{
\code{make_projection()} outputs the agent network or the symbolic
network corresponding to a survey, i.e. the row or column projection.
}
\examples{
make_check(df, "agent")
make_projection(df, "agent")
make_projection(df, "agent")
make_projection(df, "agent")
make_projection(df, "agent", threshold_method="target_lcc")
}
