% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveygraph.R
\docType{package}
\name{surveygraph}
\alias{surveygraph-package}
\alias{surveygraph}
\title{surveygraph: network representations of attitudes}
\description{
This page is a work in progress.
}
\details{
The surveygraph package functions can be summarised as follows.
}
\section{reading functions}{

The reading functions import survey datasets to R so they can be
passed to C++ routines. A bunch of file formats need to be accounted for.
}

\section{network generating functions}{

The network functions are implemented in C++.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/surveygraph/surveygraphr}
  \item \url{https://surveygraph.ie/}
  \item Report bugs at \url{https://github.com/surveygraph/surveygraphr/issues}
}

}
\author{
\strong{Maintainer}: Samuel Unicomb \email{samuelunicomb@gmail.com}

Authors:
\itemize{
  \item Ana Jovancevic
  \item Alejandro Dinkelberg
  \item Pádraig MacCarron
  \item David O'Sullivan
  \item Paul Maher
  \item Mike Quayle
}

}
