% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveygraphr.R
\docType{package}
\name{surveygraphr}
\alias{surveygraphr}
\title{surveygraphr: network representations of attitudes}
\description{
The surveygraphr package provides the following functions....
}
\section{reading functions}{

The reading functions import survey datasets to R so they can be 
passed to C++ routines. A bunch of file formats need to be accounted for.
}

\section{network generating functions}{

The network functions are implemented in C++, and blah blah blah.
}

