% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{make_projection}
\alias{make_projection}
\title{Outputs the survey projection onto the agent or symbolic layer}
\usage{
make_projection(
  data,
  layer,
  threshold_method = NULL,
  method_value = NULL,
  centre = NULL,
  similarity_metric = NULL
)
}
\arguments{
\item{data}{A data frame corresponding to a survey}

\item{layer}{A string flag that specifies which layer to project, either agent or symbolic}

\item{threshold_method}{A string flag that specifies how to threshold the network}

\item{method_value}{A utility numerical value that we interpret according to the threshold_method chosen}
}
\value{
A data.frame corresponding to the edge list of the specified network. It contains
three columns, the first two being the indices of nodes adjacent to an edge, and the 
third column its weight.
}
\description{
`make_projection()` outputs the agent network or the symbolic
network corresponding to a survey, i.e. the row or column projection.
}
\details{
add some details here
}
\examples{
make_check(df, "agent")
make_projection(df, "agent")
make_projection(df, "agent")
make_projection(df, "agent")
make_projection(df, "agent", threshold_method="target_lcc")
}
