% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic-data.R
\name{make_synthetic_data}
\alias{make_synthetic_data}
\title{Outputs a synthetic survey using a simple model}
\usage{
make_synthetic_data(nrow = 200, ncol = 15, minority = 0.5, polarisation = 0)
}
\arguments{
\item{nrow}{The number of rows in the survey}

\item{ncol}{The number of columns in the survey}

\item{minority}{The fraction of nodes in the smaller of the polariseed group}

\item{polarisation}{The degree of polarisation among agents}
}
\value{
A data frame corresponding to a survey.
}
\description{
\code{make_synthetic_data()} outputs a synthetic survey.
}
\examples{
S <- make_synthetic_data(20, 5)
}
