% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveygraphr.R
\docType{package}
\name{surveygraphr}
\alias{surveygraphr}
\title{surveygraphr: a package for visualing correlations in survey datasets}
\description{
The surveygraphr package provides the following functions....
}
\section{reading functions}{

The reading functions import survey datasets to R so they can be 
passed to C++ routines. A bunch of file formats need to be accounted for.
}

\section{network generating functions}{

The network functions are implemented in C++, and blah blah blah.
}

\section{writing functions}{

The writing functions are implemented in both R and C++ as needed, and
produce files then will go on to be read by graph visualisation libraries.
}

